adding in flask migrate and breaking out the models file
