+++ web/requirements.txt
@@ -1,2 +1,3 @@
 flask==1.1.2
 flask_sqlalchemy==2.4.1
+Flask-Migrate==2.5.3
