+++ web/web/__init__.py
@@ -1,4 +1,6 @@
 from flask import Flask
+from flask_migrate import Migrate
+
 app = Flask(__name__)
 
 import web.api.v1.api
