+++ web/web/models/meter.py
@@ -13,67 +13,6 @@ dbpass = os.environ.get('dbpass', '')
 #Instatiate a SQLAlchemy object
 db = SQLAlchemy()
 
-
-class Rate(db.Model):
-
-    __tablename__ = 'rates'
-
-    rate_id = db.Column(db.Integer, autoincrement=True, primary_key=True, nullable=False)
-    description = db.Column(db.Text, nullable=False)
-
-    #many-to-one intervals per rate
-    intervals = db.relationship('Interval', backref=db.backref('rate'))
-
-    def __repr__(self):
-        return f'<Rate rate_id={self.rate_id} description={self.description}>'
-
-
-class Address(db.Model):
-
-    __tablename__ = 'addresses'
-
-    address_id = db.Column(db.Integer, autoincrement=True, primary_key=True, nullable=False)
-    address = db.Column(db.String(100), nullable=False)
-    address2 = db.Column(db.String(64))
-    district = db.Column(db.String(64))
-    city = db.Column(db.String(100), nullable=False)
-    country = db.Column(db.String(100), nullable=False)
-    postal_code = db.Column(db.String(64), nullable=False)
-    phone = db.Column(db.String(64))
-    last_update = db.Column(TIMESTAMP, nullable=False)
-
-    def __repr__(self):
-        return f'<Address address_id={self.address_id} address={self.address} postal_code={self.postal_code}>'
-
-
-class ServiceLocation(db.Model):
-
-    __tablename__ = 'service_locations'
-
-    service_location_id = db.Column(db.String(64), primary_key=True, nullable=False)
-    address_id = db.Column(db.Integer, db.ForeignKey('addresses.address_id'), nullable=False)
-    map_location = db.Column(db.String(64), nullable=False)
-
-    #one-to-one service location per address
-    address = db.relationship('Address', backref=db.backref('service_locations'), uselist=False)
-
-    def __repr__(self):
-        return f'<ServiceLocation service_location_id={self.service_location_id} address_id={self.address_id}>'
-
-
-class Utility(db.Model):
-    
-    __tablename__ = 'utilities'
-
-    utility_id = db.Column(db.Integer, autoincrement=True, primary_key=True, nullable=False)
-    name = db.Column(db.String(64), nullable=False)
-    subscription_start = db.Column(TIMESTAMP, nullable=False)
-    subscription_end = db.Column(TIMESTAMP, nullable=False)
-
-    def __repr__(self):
-        return f'<Utility utility_id={self.utility_id} name={self.name}>'
-
-
 #enum fields for meter_type column in meter
 class MeterType(enum.Enum):
 
@@ -124,82 +63,6 @@ class Meter(db.Model):
     def __repr__(self):
 
         return f'<Meter meter_id={self.meter_id} is_active={self.is_active}>'
-
-
-class Channel(db.Model):
-
-    __tablename__ = 'channels'
-
-    channel_id = db.Column(db.Integer, autoincrement=True, primary_key=True, nullable=False)
-    meter_id = db.Column(db.String(64), nullable=False)
-    utility_id = db.Column(db.Integer, nullable=False)
-    service_location_id = db.Column(db.String(64), nullable=False)
-    setting = db.Column(db.Integer, nullable = False) 
-    channel_type = db.Column(db.String(64), nullable=False)
-
-    #composite foreign key to meter
-    __table_args__ = (ForeignKeyConstraint([meter_id, utility_id, service_location_id],
-                                           [Meter.meter_id, Meter.utility_id, Meter.service_location_id]), 
-                                           {})
-
-    #many-to-one channels per meter
-    meter = db.relationship('Meter', backref=db.backref('channels'))
-
-    def __repr__(self):
-
-        return f'<Channel channel_id={self.channel_id} meter_id={self.meter_id} channel_type={self.channel_type}>'
-
-
-#enum fields for status column in interval
-class Status(enum.Enum):
-
-    #what values/names for status?
-    one = "valid"
-
-
-class Interval(db.Model):
-
-    __tablename__ = 'intervals'
-    
-    interval_id = db.Column(db.Integer, autoincrement=True, primary_key=True, nullable=False)
-    meter_id = db.Column(db.String(64), nullable=False)
-    utility_id = db.Column(db.Integer, nullable=False)
-    service_location_id = db.Column(db.String(64), nullable=False)
-    rate_id = db.Column(db.Integer, db.ForeignKey('rates.rate_id'), nullable=False) 
-    status = db.Column(db.Enum(Status), nullable=False)
-    start_time = db.Column(TIMESTAMP, nullable=False)
-    end_time = db.Column(TIMESTAMP, nullable=False)
-    value = db.Column(db.Float, nullable=False)
-
-    #composite foreign key to meter
-    __table_args__ = (ForeignKeyConstraint([meter_id, utility_id, service_location_id],
-                                           [Meter.meter_id, Meter.utility_id, Meter.service_location_id]), 
-                                           {})
-
-    #many-to-one intervals per meter
-    meter = db.relationship('Meter', backref=db.backref('intervals'))
-    
-    #many-to-one intervals per rate
-    rate = db.relationship('Rate', backref=db.backref('intervals'))
-
-    @staticmethod
-    def get_interval_coverage(interval_id_list):
-        '''Takes in list of interval ids, 
-            returns list of tuples for start and end times in ISO8601 format'''
-        
-        #improve for error handling (id doesn't exist in list)
-
-        selected_intervals = Interval.query.filter(Interval.interval_id.in_(interval_id_list)).all()
-        start_end_tuples_list = []
-
-        for interval in selected_intervals:
-            start_end_tuples_list.append((interval.start_time, interval.end_time))
-
-        return start_end_tuples_list
-
-    def __repr__(self):
-
-        return f'<Interval interval_id={self.interval_id} meter_id={self.meter_id} end_time={self.end_time} value={self.value}>'
 
 
 #local connection - switch to rds to deploy
